Add basic pagination for table. Fetch HOC. Folder restructure
