+++ package.json
@@ -43,7 +43,7 @@
     "@lumino/widgets": "^1.14.0",
     "@types/lodash": "^4.14.167",
     "@types/react-router-dom": "^5.1.6",
-    "@types/react-table": "^7.0.26",
+    "@types/react-table": "^7.0.27",
     "@types/react-tabs": "^2.3.2",
     "@types/styled-components": "^5.1.5",
     "axios": "^0.21.1",
@@ -55,7 +55,7 @@
     "prop-types": "^15.7.2",
     "query-string": "^6.13.8",
     "react-router-dom": "^5.2.0",
-    "react-table": "^7.6.2",
+    "react-table": "^7.6.3",
     "react-tabs": "^3.1.2",
     "react-tooltip": "^4.2.11",
     "styled-components": "^5.2.1",
