+++ src/utils/constants.ts
@@ -6,3 +6,7 @@ export enum API_STATUSES {
 
 //export const BACKEND_HOST = 'http://localhost:8000';
 export const BACKEND_HOST = 'https://beta.mamba.pm';
+
+export const KB_SIZE = 1024;
+export const MB_SIZE = 1024 * KB_SIZE;
+export const GB_SIZE = 1024 * MB_SIZE;
