+++ style/about.css
@@ -1,30 +1,34 @@
 .about-page {
   width: 100%;
-  height: 100%;
-}
-
-.about {
-  width: 100%;
-  height: 100%;
+  min-height: calc(100vh - var(--jp-private-topbar-height));
+  display: flex;
+  justify-content: center;
   background: url(img/mamba.jpg) no-repeat center center fixed;
-  -webkit-background-size: cover;
-  -moz-background-size: cover;
-  -o-background-size: cover;
   background-size: cover;
 }
 
-.about .description {
-  width: 600px;
-  height: 500px;
-  position: absolute;
-  top: 100px;
-  left: 100px;
+.about {
+  min-width: 70%;
+  padding: 20px;
+  margin: 0 auto;
 }
 
-.about .description h1 {
+.about hr {
+  border-color: var(--jp-quetz-primary-color);
+  margin-bottom: 30px;
+}
+
+.about h1 {
   color: var(--jp-quetz-primary-color);
 }
 
-.about .description p {
+.about p {
+  line-height: 30px;
+  text-shadow: 0 0 5px black;
+}
+
+.about-list {
+  margin: 0;
   padding-left: 20px;
+  list-style: none;
 }
