+++ style/download.css
@@ -1,57 +1,48 @@
 .download-page {
   width: 100%;
-  height: 100%;
+  min-height: calc(100vh - var(--jp-private-topbar-height));
+  display: flex;
+  justify-content: center;
+  background: url(img/mamba.jpg) no-repeat center center fixed;
+  background-size: cover;
 }
 
 .download {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center center;
-  background: url(img/mamba.jpg) no-repeat center center fixed;
-  -webkit-background-size: cover;
-  -moz-background-size: cover;
-  -o-background-size: cover;
-  background-size: cover;
-}
-
-.download .description {
-  width: 70%;
-  height: 70%;
-  margin: auto;
+  min-width: 70%;
+  padding: 20px;
+  margin: 0 auto;
   color: var(--jp-quetz-primary-color);
 }
 
-.download .description hr {
+.download hr {
   border-color: var(--jp-quetz-primary-color);
+  margin-bottom: 30px;
 }
 
-.download .description p {
-  padding-left: 20px;
+.download .paragraph {
+  line-height: 30px;
 }
 
-.download .description .windows {
+.download-button {
   width: 100px;
-  height: 125px;
-  margin: 50px;
-  display: flex;
-  flex-direction: column;
+  height: 120px;
+  margin: 10px 30px;
+  border: 0;
   background-color: var(--jp-quetz-primary-color);
-  border: solid var(--jp-quetz-primary-color);
   border-radius: 10px;
   color: var(--jp-layout-color1);
+  font-weight: bold;
 }
 
-.download .description .windows .logo {
+.download-logo {
   width: 50px;
   height: 50px;
-  margin: auto;
+  display: block;
+  margin: 0 auto;
   background: url(img/windows-logo.svg) no-repeat 50% 50%;
   background-size: cover;
-  color: var(--jp-layout-color1);
-  fill: var(--jp-layout-color1);
 }
 
-.download .description .windows .label {
-  margin: auto;
+.download-label {
+  margin: 10px auto 0;
 }
